backend/requirements.txt
Flask
python-dotenv
Flask-Bcrypt
mercadopago